Remember bottomBar tab using compose-router
The tab to be selected is sent by the returning screen using navTo,
which is called when the up button on the topBar is clicked. So really,
the tab isn't "remembered", it's just being passed back.
This causes a problem: When compose-router handles up navigation itself,
navTo isn't called, and the tab is reset to Routines.